# public/openapi/apis.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "isrvd API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API 服务"
    }
  ],
  "paths": {
...
    "/filer/chmod": {
      "put": {
        "operationId": "put_filer_chmod",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/filerChmodBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "修改文件权限",
        "tags": [
          "filer"
        ]
      }
    },
    "/filer/dir": {
      "post": {
        "operationId": "post_filer_dir",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/filer.filerFileMkdir"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "创建目录",
        "tags": [
          "filer"
        ]
      }
    },
    "/filer/dir-size": {
      "get": {
        "operationId": "get_filer_dir_size",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "计算目录大小",
        "tags": [
          "filer"
        ]
      }
    },
    "/filer/download": {
      "get": {
        "operationId": "get_filer_download",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "inline",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "下载文件",
        "tags": [
          "filer"
        ]
      }
    },
    "/filer/file": {
      "delete": {
        "operationId": "delete_filer_file",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "删除文件或目录",
        "tags": [
          "filer"
        ]
      },
      "get": {
        "operationId": "get_filer_file",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "读取文件内容",
        "tags": [
          "filer"
        ]
      },
      "post": {
        "operationId": "post_filer_file",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/filerContentBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "创建文件",
        "tags": [
          "filer"
        ]
      },
      "put": {
        "operationId": "put_filer_file",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/filerContentBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "保存文件内容",
        "tags": [
          "filer"
        ]
      }
    },
    "/filer/files": {
      "get": {
        "operationId": "get_filer_files",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "查询目录文件列表",
        "tags": [
          "filer"
        ]
      }
    },
    "/filer/rename": {
      "post": {
        "operationId": "post_filer_rename",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/filerRenameBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "重命名文件或目录",
        "tags": [
          "filer"
        ]
      }
    },
    "/filer/unzip": {
      "post": {
        "operationId": "post_filer_unzip",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/filerUnzipBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "解压文件",
        "tags": [
          "filer"
        ]
      }
    },
    "/filer/upload": {
      "post": {
        "operationId": "post_filer_upload",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "上传文件",
        "tags": [
          "filer"
        ]
      }
    },
    "/filer/zip": {
      "post": {
        "operationId": "post_filer_zip",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/filerPathQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "压缩文件或目录",
        "tags": [
          "filer"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIResponse": {
        "properties": {
          "message": {
            "type": "string"
          },
          "payload": {},
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "filer.filerFileMkdir": {
        "properties": {
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "type": "object"
      },
      "filerChmodBody": {
        "properties": {
          "mode": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "mode"
        ],
        "type": "object"
      },
      "filerContentBody": {
        "properties": {
          "content": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "content"
        ],
        "type": "object"
      },
      "filerPathQuery": {
        "properties": {
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "type": "object"
      },
      "filerRenameBody": {
        "properties": {
          "path": {
            "type": "string"
          },
          "target": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "target"
        ],
        "type": "object"
      },
      "filerUnzipBody": {
        "properties": {
          "path": {
            "type": "string"
          },
          "targetDir": {
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  }
}